+++ src/filters/axios-exceptions/index.ts
@@ -3,12 +3,15 @@ import {
   ArgumentsHost,
   HttpStatus,
   ExceptionFilter,
+  Logger,
 } from '@nestjs/common';
 import { AxiosError } from 'axios';
 import { Response } from 'express';
 
 @Catch(AxiosError)
 export class AxiosExceptionsFilter implements ExceptionFilter {
+  constructor(private readonly logger: Logger) {}
+
   catch(exception: AxiosError, host: ArgumentsHost) {
     const ctx = host.switchToHttp();
     const expressResponse = ctx.getResponse<Response>();
@@ -23,6 +26,8 @@ export class AxiosExceptionsFilter implements ExceptionFilter {
       firstLetterUpperCase(
         HttpStatus[statusCode].replace(/_/g, ' ').toLocaleLowerCase(),
       ) || 'Internal Server Error';
+
+    this.logger.error(message, exception.stack, 'AxiosError');
 
     expressResponse.status(statusCode).json({
       statusCode,