+++ src/modules/movies/index.ts
@@ -1,4 +1,4 @@
-import { Module } from '@nestjs/common';
+import { Logger, Module } from '@nestjs/common';
 import { MoviesController } from 'src/controllers/movies';
 import { MoviesService } from 'src/services/movies';
 import { StarwarsRepository } from 'src/repositories/starwars';
@@ -8,6 +8,12 @@ import { Database } from 'src/infrastructure/database';
 @Module({
   imports: [],
   controllers: [MoviesController],
-  providers: [Database, MoviesRepository, MoviesService, StarwarsRepository],
+  providers: [
+    Database,
+    MoviesRepository,
+    MoviesService,
+    StarwarsRepository,
+    Logger,
+  ],
 })
 export class MoviesModule {}