+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Enhace error logging ([#8](https://github.com/nahuelsoma/movies-api/pull/8))
 - Enhace error handling ([#7](https://github.com/nahuelsoma/movies-api/pull/7))
 - Add user sign up, login and endpoint restriction by roles ([#5](https://github.com/nahuelsoma/movies-api/pull/5))
 - Add delete movie method ([#4](https://github.com/nahuelsoma/movies-api/pull/4))